index.html - Don't show quite as much of the innards
You can still see all this stuff on Github, but it feels
weird to give this content when some goes trawling the URLs.
The only supported/maintained interface is `stable.php`; any
other details/end-points could change at any time, so best
not to encourage people to rely on them.